+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Update README to note Node version restrictions.
 - Maintain the value of the channel color in HSV when applying intensity (3D + 2D)
 - Fix documentation for React Components.
+- Fix channel stats for thin 3D volumes in Avivator.
+- Better checking for viewable volumes at certain resolutions.
 - Don't do RGB-HSV conversion on shaders.  Instead, scale RGB linearly.
 - Consistent shader floats `n.0` -> `n.`
 - Change API for all public props
@@ -19,6 +21,7 @@
   - `channelIsOn` -> `active`
   - `loaderSelection` -> `selections`
   - `MAX_SLIDERS_AND_CHANNELS` -> `MAX_CHANNELS`
+
 ## 0.10.5
 
 ### Added